Now generating Kotlin Binders, removed "nimbus" module prefix
