+++ platforms/android/annotations/gradle.properties
