+++ platforms/android/annotations/src/main/java/com/salesforce/nimbus/PluginOptions.kt
