+++ platforms/android/bridge-webview/.gitignore
