+++ platforms/android/bridge-webview/gradle.properties
@@ -1,0 +1,3 @@
+POM_NAME=Nimbus Bridge WebView
+POM_PACKAGING=aar
+kapt.use.worker.api=true
