+++ platforms/android/bridge-webview/proguard-rules.pro
