+++ platforms/android/bridge-webview/src/androidTest/AndroidManifest.xml
