+++ platforms/android/bridge-webview/src/androidTest/res/layout/activity_web_view.xml
