+++ platforms/android/bridge-webview/src/main/AndroidManifest.xml
