+++ platforms/android/bridge-webview/src/main/res/values/strings.xml
