+++ platforms/android/compiler-base/gradle.properties
@@ -1,0 +1,2 @@
+POM_NAME=Nimbus Compiler
+POM_PACKAGING=jar
