+++ platforms/android/compiler-webview/build.gradle
@@ -1,0 +1,61 @@
+import groovy.json.JsonSlurper
+
+apply plugin: 'java-library'
+apply plugin: 'kotlin'
+
+dependencies {
+    //noinspection DifferentStdlibGradleVersion
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
+    implementation project(':compiler-base')
+    implementation project(':annotations')
+    implementation 'com.squareup:kotlinpoet:1.5.0'
+    implementation 'org.jetbrains.kotlinx:kotlinx-metadata-jvm:0.1.0'
+}
+
+buildscript {
+
+    def versionFile = file('../../../lerna.json')
+    def parsedVersion = new JsonSlurper().parseText(versionFile.text)
+    project.version = parsedVersion.version.toString()
+
+    repositories {
+        mavenCentral()
+    }
+    dependencies {
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+    }
+}
+
+repositories {
+    mavenCentral()
+}
+
+compileKotlin {
+    kotlinOptions {
+        jvmTarget = "1.8"
+    }
+}
+
+compileTestKotlin {
+    kotlinOptions {
+        jvmTarget = "1.8"
+    }
+}
+
+task sourcesJar(type: Jar, dependsOn:classes) {
+    archiveClassifier = 'sources'
+    from sourceSets.main.allSource
+}
+
+task javadocJar(type: Jar, dependsOn:javadoc) {
+    archiveClassifier = 'javadoc'
+    from javadoc.destinationDir
+}
+
+artifacts {
+    archives javadocJar
+    archives sourcesJar
+}
+
+apply from: rootProject.file('gradle/lint.gradle')
+apply from: rootProject.file('gradle/publishing.gradle')
