+++ platforms/android/compiler-webview/gradle.properties
