+++ platforms/android/compiler-webview/src/main/java/com/salesforce/nimbus/bridge/webview/compiler/WebViewBinderGenerator.kt
@@ -1,0 +1,9 @@
+package com.salesforce.nimbus.bridge.webview.compiler
+
+import com.salesforce.nimbus.compiler.BinderGenerator
+import com.squareup.kotlinpoet.ClassName
+
+class WebViewBinderGenerator : BinderGenerator() {
+    override val javascriptEngine = ClassName("android.webkit", "WebView")
+    override val functionAnnotationClassName = ClassName("android.webkit", "JavascriptInterface")
+}
