+++ platforms/android/core/src/main/AndroidManifest.xml
