+++ platforms/android/core/src/main/java/com/salesforce/nimbus/Bridge.kt
