+++ platforms/android/core/src/main/java/com/salesforce/nimbus/JSONSerializable.kt
