+++ platforms/android/core/src/main/java/com/salesforce/nimbus/Plugin.kt
