--- platforms/android/nimbus-bridge-webview/gradle.properties
@@ -1,2 +1,0 @@
-POM_NAME=Nimbus Bridge WebView
-POM_PACKAGING=aar
