--- platforms/android/nimbus-core/build.gradle
@@ -1,60 +1,0 @@
-import groovy.json.JsonSlurper
-
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-apply plugin: 'org.jetbrains.dokka'
-
-android {
-    compileSdkVersion 29
-
-    def versionFile = file('../../../lerna.json')
-    def parsedVersion = new JsonSlurper().parseText(versionFile.text)
-    project.version = parsedVersion.version.toString()
-
-    defaultConfig {
-        minSdkVersion 21
-        targetSdkVersion 29
-        versionCode 1
-        versionName version
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-            testCoverageEnabled false
-        }
-        debug {
-            testCoverageEnabled true
-        }
-    }
-
-    testOptions {
-        unitTests.returnDefaultValues = true
-    }
-}
-
-dependencies {
-    api project(":nimbus-annotations")
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
-    testImplementation 'io.kotlintest:kotlintest-runner-junit4:3.1.5'
-    testImplementation 'org.json:json:20190722'
-}
-
-task sourcesJar(type: Jar) {
-    archiveClassifier = 'sources'
-    from android.sourceSets.main.java.srcDirs
-}
-
-artifacts {
-    archives sourcesJar
-}
-
-dokka {
-    outputFormat = 'html'
-    outputDirectory = "$buildDir/dokka"
-}
-
-apply from: rootProject.file('gradle/lint.gradle')
-
