+++ platforms/android/settings.gradle
@@ -1,2 +1,2 @@
 rootProject.name = 'Nimbus'
-include ':demo-app', ':nimbus-bridge-webview', ':nimbus-annotations', ':nimbus-compiler-webview', ':nimbusjs', ':nimbus-core'
+include ':demo-app', ':bridge-webview', ':annotations', ':compiler-webview', ':nimbusjs', ':core', ':compiler-base'
